packages/database-api-endpoint/database-api-endpoint/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.2; python_full_version >= '3.6.2'
atpublic==3.1.1; python_version >= '3.7'
click==8.1.3; python_version >= '3.7'
commonmark==0.9.1
fastapi==0.85.1
h11==0.14.0; python_version >= '3.7'
ibis-framework[postgres]==3.2.0
idna==3.4; python_version >= '3.5'
multipledispatch==0.6.0
numpy==1.23.4; python_version >= '3.8'
packaging==21.3; python_version >= '3.6'
pandas==1.5.1; python_version >= '3.8'
parsy==2.0; python_version >= '3.6'
psycopg2==2.9.4
pydantic==1.10.2; python_version >= '3.7'
pygments==2.13.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
pytz==2022.5
regex==2022.9.13; python_version >= '3.6'
rich==12.6.0; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==1.4.42
sqlglot==6.3.5
starlette==0.20.4; python_version >= '3.7'
toolz==0.12.0; python_version >= '3.5'
typing-extensions==4.4.0; python_version >= '3.7'
uvicorn==0.19.0
